+++ lib/libutee/sub.mk
@@ -3,6 +3,7 @@ global-incdirs-y += include
 srcs-y += abort.c
 srcs-y += trace_ext.c
 srcs-y += assert.c
+srcs-y += tee_uuid_from_str.c
 ifneq ($(sm),ldelf)
 srcs-y += tee_api_property.c
 srcs-y += base64.c
